Refactor net package and add DoPost helper
- Add net.DoPost generic helper for mTLS POST with request marshaling
- Refactor secret and acl packages to use net.DoPost, removing boilerplate
- Consolidate net package files:
  - Merge header.go into types.go
  - Merge name.go, policy.go, permission.go into validate.go
  - Merge fallback.go, unauthorized.go into respond.go
  - Merge parse.go into guard.go
- [BREAKING] Move Source() from net/spiffe.go to spiffe.SourceOrDie(ctx)
  with context parameter for timeout control
