+++ api/internal/impl/secret/delete.go
@@ -6,7 +6,6 @@ package secret
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
 
@@ -43,20 +42,10 @@ func Delete(
 	source *workloadapi.X509Source,
 	path string, versions []int,
 ) *sdkErrors.SDKError {
-	if source == nil {
-		return sdkErrors.ErrSPIFFENilX509Source.Clone()
-	}
-
 	r := reqres.SecretDeleteRequest{Path: path, Versions: versions}
 
-	mr, marshalErr := json.Marshal(r)
-	if marshalErr != nil {
-		failErr := sdkErrors.ErrDataMarshalFailure.Wrap(marshalErr)
-		failErr.Msg = "problem generating the payload"
-		return failErr
-	}
-
-	_, postErr := net.PostAndUnmarshal[reqres.SecretDeleteResponse](
-		ctx, source, url.SecretDelete(), mr)
-	return postErr
+	_, err := net.DoPost[reqres.SecretDeleteResponse](
+		ctx, source, url.SecretDelete(), r,
+	)
+	return err
 }
