+++ config/fs/validate.go
