--- net/header.go
@@ -1,11 +1,0 @@
-//    \\ SPIKE: Secure your secrets with SPIFFE. — https://spike.ist/
-//  \\\\ Copyright 2024-present SPIKE contributors.
-// \\\\\\ SPDX-License-Identifier: Apache-2.0
-
-package net
-
-type ContentType string
-
-var ContentTypeJSON ContentType = "application/json"
-var ContentTypePlain ContentType = "text/plain"
-var ContentTypeOctetStream ContentType = "application/octet-stream"
